Artificialinteligence/imagecaptioning/requirements.txt
flask
torch
torchvision
pytorch-lightning
sentencepiece
timm
nltk
scikit-learn
Pillow
